fix(bot-mode): stop isolated turns on teardown
